rename rename component in rename container
